[Oztechan/CCC#1863] Move CoroutinesUtil and toStandardDigits into :client:core:viewmodel
