+++ client/core/viewmodel/src/commonMain/kotlin/com/oztechan/ccc/client/viewmodel/util/CoroutineUtil.kt
@@ -1,16 +1,16 @@
-package com.oztechan.ccc.client.util
+package com.oztechan.ccc.client.viewmodel.util
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 
-internal fun CoroutineScope.launchIgnored(function: suspend () -> Unit) {
+fun CoroutineScope.launchIgnored(function: suspend () -> Unit) {
     launch {
         function()
     }
 }
 
-internal inline fun <T> MutableStateFlow<T>.update(function: T.() -> T) {
+inline fun <T> MutableStateFlow<T>.update(function: T.() -> T) {
     update { function(value) }
 }
