+++ client/core/viewmodel/src/commonTest/kotlin/com/oztechan/ccc/client/core/viewmodel/util/FormatUtilTest.kt
@@ -1,6 +1,7 @@
 package com.oztechan.ccc.client.core.viewmodel.util
 
 import com.oztechan.ccc.client.viewmodel.util.getFormatted
+import com.oztechan.ccc.client.viewmodel.util.toStandardDigits
 import com.oztechan.ccc.client.viewmodel.util.toSupportedCharacters
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -25,5 +26,31 @@ class FormatUtilTest {
         assertEquals("0.000000001", actualDouble5.getFormatted(3))
         val actualDouble6 = 0.0000000001
         assertEquals("0", actualDouble6.getFormatted(3))
+    }
+
+    @Test
+    fun toStandardDigits() {
+        // https://en.wikipedia.org/w/index.php?title=Hindu%E2%80%93Arabic_numeral_system
+        listOf(
+            "०१२३४५६७८९",
+            "૦૧૨૩૪૫૬૭૮૯",
+            "੦੧੨੩੪੫੬੭੮੯",
+            "༠༡༢༣༤༥༦༧༨༩",
+            "০১২৩৪৫৬৭৮৯",
+            "೦೧೨೩೪೫೬೭೮೯",
+            "୦୧୨୩୪୫୬୭୮୯",
+            "൦൧൨൩൪൫൬൭൮൯",
+            "౦౧౨౩౪౫౬౭౮౯",
+            "០១២៣៤៥៦៧៨៩",
+            "๐๑๒๓๔๕๖๗๘๙",
+            "໐໑໒໓໔໕໖໗໘໙",
+            "၀၁၂၃၄၅၆၇၈၉",
+            "٠١٢٣٤٥٦٧٨٩",
+            "۰۱۲۳۴۵۶۷۸۹",
+            "۰۱۲۳۴۵۶۷۸۹",
+            "᠐᠑᠒᠓᠔᠕᠖᠗᠘᠙"
+        ).forEach {
+            assertEquals("0123456789", it.toStandardDigits(), "actual string $it")
+        }
     }
 }
