+++ client/src/commonMain/kotlin/com/oztechan/ccc/client/viewmodel/premium/PremiumViewModel.kt
@@ -13,8 +13,8 @@ import com.oztechan.ccc.client.model.PremiumData
 import com.oztechan.ccc.client.model.PremiumType
 import com.oztechan.ccc.client.storage.app.AppStorage
 import com.oztechan.ccc.client.util.calculatePremiumEnd
-import com.oztechan.ccc.client.util.launchIgnored
-import com.oztechan.ccc.client.util.update
+import com.oztechan.ccc.client.viewmodel.util.launchIgnored
+import com.oztechan.ccc.client.viewmodel.util.update
 import com.oztechan.ccc.client.viewmodel.BaseData
 import com.oztechan.ccc.client.viewmodel.BaseSEEDViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
